requirements.txt
boto3==1.16.44
pytest==7.1.3
python-dotenv==0.21.0